# src/backend/api-spec.yaml
openapi: 3.0.0
info:
  title: Hospital Network Management System API
  description: |
    API specification for the Hospital Network Management System,
    enabling secure interoperability between hospital systems.
  version: 1.0.0
servers:
  - url: https://api.hospital-network.org/v1
    description: Production API server
  - url: https://staging-api.hospital-network.org/v1
    description: Staging API server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Medication:
      type: object
      required:
        - rxnormCode
        - name
        - form
        - strength
      properties:
        medicationId:
          type: string
          format: uuid
          readOnly: true
        rxnormCode:
          type: string
          description: RxNorm standard code
        snomedCtCode:
          type: string
          description: SNOMED CT code if available
          nullable: true
        name:
          type: string
        form:
          type: string
          enum: [tablet, capsule, liquid, injection, patch, inhaler, other]
        strength:
          type: string
        manufacturer:
          type: string
          nullable: true
        barcodeGs1:
          type: string
          description: GS1 standard barcode
          nullable: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    PatientMedication:
      type: object
      required:
        - patientId
        - medicationId
        - hospitalId
        - prescriberId
        - dosage
        - frequency
        - route
        - prescribedDate
        - startDate
      properties:
        patientMedicationId:
          type: string
          format: uuid
          readOnly: true
        patientId:
          type: string
          format: uuid
        medicationId:
          type: string
          format: uuid
        hospitalId:
          type: string
          format: uuid
        prescriberId:
          type: string
          format: uuid
        dosage:
          type: string
        frequency:
          type: string
        route:
          type: string
          enum: [oral, intravenous, intramuscular, subcutaneous, topical, inhalation, other]
        prescribedDate:
          type: string
          format: date
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        instructions:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /patients/{patientId}/medications:
    post:
      summary: Add a medication to patient's record
      operationId: addPatientMedication
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientMedication'
      responses:
        '201':
          description: Medication added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientMedication'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Patient or medication not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List patient medications
      operationId: listPatientMedications
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: active
          in: query
          schema:
            type: boolean
        - name: hospitalId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of patient medications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatientMedication'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /medications:
    post:
      summary: Create a new medication
      operationId: createMedication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Medication'
      responses:
        '201':
          description: Medication created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medication'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Search medications
      operationId: searchMedications
      parameters:
        - name: rxnormCode
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: barcodeGs1
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of medications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Medication'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...